Updated Toast
I have moved Toast to a higher level, so it doesn't need to constantly be imported. I have also been expirementing with different ways of showing the notifications. For now, I have put the notification to the center of the screen. I want to find a way to customize the colors a little bit more than just having it be green.