2 Step Signup Process
I have added a new file/page, which is the second step to the sign in process. The first step asks for an email and then sends a confirmation email. Then, the second page asks for a password. This makes sure only valid emails are used to make accounts. Also, I changed the button on the first page to say 'send confirmation' from 'create account'. This was because at that step, they weren't actually making an account, but just sending a confirmation email.