Login and Signup Changes
I have built out the login page partially. It inlcudes a password and email fields as well as changes to the bottom text that will link you to the signup page. I also made changes across both the signup and login page. There are links below the form. Prior, these were <a> tags. But, I switched these to <Link> tags.