Created User Previous Orders Page
I have created a new page for the user. This new page will include things like the user's previous orders, products that they have previously looked at, etc. Also, I want to use a feature from AliExpress that has a section for the user's orders that have not yet been reviewed. This will increase the chances of the user leaving a review, and there could be some possible incentive for leaving a review as well.