Added Link Active
On the user side nav in their dashboard, there is now a method to check if the current path matches the path of the side nav link. This way, when you switch pages the syling of the side nav links also change to indicate the current page that you are on.