Added Admin Side Nav
There is now a side navigation for admin. This has multiple different routes and tools. Also, at the bottom there is an option to change the password. This will take you to the user dashboard. However, if you are an admin, all the other user related things like history, open orders, etc. will not appear for the admin.