Added Front End Category Creation
You can now create categories from the admin dashboard. The input field that I added yesterday is now functioning. After submitting a category, you will get a message saying that it was created. To avoid duplicate categories, if you put the same category in multiple times, it will only be created once. There was also an issue where the user's token was failing to verify on the backend. This was a simple mistake, and was fixed by switching around two variables in the createCategory method.