Added Loading
I have added a loading animation to the category creation form. This is far from what I wanted it to be. I tried making a pulsating input form, however the loading was so quick that it just didn't look right, so I simplified it down a bit more. But, I guess quick loading is far better than having to play a loading animation.