Added Deletion and Listing
On the category page of the admin dashboard, you can now see a list of the current categories. There are also icons next to each category that will allow you to update or delete the category. If you choose to delete a category, a message will appear asking if you want to confirm. Then it will be deleted.