Added Search and Stat
You can now search through the categories on the categories page of the dashboard. Also, I added the search field on the top next to the create category field. Another thing I added was the categories stat card. This card shows the number of categories that are there. This may be modified or duplicated to include other future category attributes, such as hot categories.