Subcategories and filtering
This was a massive addition and there are likely to be issues. However, I believe most of the issues have been resolved. On the subcategory page, you can now create new subcategories. Just select the parent category, and type the subcategory name and it will be created. The search field, category area, and stat card are all still involved with the regular categories and still need to be changed for subcategories. The last change was filtering. I wanted the select box to be sorted by alphabetical order, so you could easily find the parent category. To do this, I modified the API to have a filter option. This fully works and allows you to sort the categories by date created, or by alphabetical order. This small change will make it easier to create sub categories and sort through a long list of categories.