Listing Subcategories
I have added a function that gets the subcategories. This was very similar to getting the categories. I also changed a lot of the things to make the subcategory page more independent of the regular categories page. It will now list out the subcategories, where the regular categories were listed. The search field also now works and will search through the sub categories. The last thing that needed fixed was the stat card in the top right. It will now display the number of subcategories. Finally, I also recalled the load subcategories function whenever one is deleted or added. This will make sure things are updated whenever something happens, so the user doesn't need to refresh the page.