Updated Navs
I noticed an issue with the admin nav. This issue was when the category was being hovered over and it looked clickable, but when you clicked it wouldn't work. It would only seem to work over or very close to the text. This has since been fixed, to include the link on the entire list item rather than just the text.