Image Preview
After much time, I have figured out how to show an image preview. Since the images are not being automatically uploaded when selected, I saved the image data to its own variable in the state. Then, it will go through the image data and insert new images into the html. There were numerous issues with implementing this. There were so many issues trying to access the image data, and calling certain functions at certain times. Also, if you reselected the images, it wasn't removing the previous images and redisplaying them. But, I found a way to remove the children. The final issue I encountered was that the images at the end were not being uploaded when you created a product. This has since been fixed.