List Products, Style, MASSIVE BUG FIX
I have added a method to get a list of the products by passing in an amount. I also went through all the admin pages and changed the height. Finally, there was a massive problem where the image, subCategories, and subsidiaryBrands were not saving to the database at all. I have found a work around and now it appears to be working.