Added Image + Style Improvements
I have aded a default image in the event that the images don't load or that there is somehow a way that a product is created without them. In this case, it will just show a default image icon. Next, I have added two buttons on the top right of each product card. These buttons, when hovered add a nice effect, but also, they will allow admin to update and delete products. For users, these will be changed and show something like adding to cart or favoriting.