Sub Reset + category and brand fix
I have added to the line that sets the brand and category, when a new field is selected. Now, the previous subs will be removed and replaced with an empty array. This is to prevent one product having multiple subcategories from multiple different categories by accident. Vice versa with the brands and subsidiary brands. I also changed a line that gave a duplicate option.