Product Update (7 hours wasted)
This took forever! There were so many bugs that popped up while making this form function. The form will be populated with information about the product. The admin can make changes to the product. Then, there was a bunch of issues with the images. However, everything appears to be working now.