Bug Fixes
Fixed multiple sub selection bug. This bug would make it so that if multiple subCategories or subBrands were selected, it would not show that they were checked when you went to update them. This has since been fixed. Another bug was more of a style change. The warning has been moved below the images and a confirm box appears before you delete an image.