Hero
I have added a hero section. Since it is close to the holiday season, I have chose to set different components for different times of the year. This way, the hero components can be swapped out as needed depending on the time of year. The one that I have added is for the holdiay season.