Hero Sections
I have added another hero section for winter. This section includes text, a link to the winter catalog, and an image. I have also changed the file names on both the old and new winter hero sections. This is to show that one has text and the other has an image, depending on what the admins may want to use.