Bug Fix
There was a bug where the form would stop loading, but the processing wasn't complete. This would possible confuse the user/admins, it may also lead to multiple submissions, where it is not necessary. This has since been fixed and the form for creating a product will remain loading until it is fully finished.