Main Image Display
Now, when you are looking at the product cards, the main image is displayed. Previously, it would just select the first image in the array of images. This created a major problem in that the admin couldn't choose the image to be displayed first. Sometimes uploading in a certain order would work, and other times it woulndn't work. This way, the admin can make sure they have the right image displayed first to the user via the product card.