Section/Page Changes
As it stands, there are 3 main sections on the home page. The first section is the new arrivals. For this section, I want there to be 4 products as a sneak peak only. The user then clicks a link to another page to view up to 16 or 20 new products. This section may also include new deals, coupons, etc. The next section is the best sellers. This are will be similar to the first. There will only be 4 as a sneak peak. When they click on the best sellers page, the user will see deals, top sellers, etc. The final section is the main product section 'everyday finds'. This section will not have its own page as of now. However, it will have pagination. It will contain about 20 products, then the user can click next or previous to bounce around pages.