Bug Fix + New Product Page
I have finally fixed a dangerous bug that would screw over every product you created. This bug would use the images from a previous product and use them on a new product. I found that this was a problem with a persistent state. To fix this, I added a line that would remove or reset the state for images and main images. Lastly, I added a new product page. This product page will display information when you access it. To get to it, just click on a product card on the main home page. Right now, it is really simple and can use more functionality.